fix: Width-coerce the len() builtin on Rust and Kotlin
- Infer len(bytes|str) as UntypedInt so comparisons adopt the width
- Cast usize Rust .len() to the expected int width at emit time
- Kotlin .size gains .toUInt(); bless 2 goldens, rest unchanged

+++ tests/forge/expected/AlgorithmBytesEqual.kt
@@ -19,7 +19,7 @@ fun bytesEqual(a: ByteArray, b: ByteArray): Boolean {
         return false
     }
     var i: UInt = 0.toUInt()
-    while (i < (a).size) {
+    while (i < (a).size.toUInt()) {
         if (a[i.toInt()].toUByte() != b[i.toInt()].toUByte()) {
             return false
         }
